[AutoNode] Move validator activation/deactivation to validator.py
