+++ AutoNode/node.py
@@ -10,7 +10,6 @@ import requests
 
 from pyhmy import (
     Typgpy,
-    cli
 )
 
 from .common import (
@@ -22,14 +21,9 @@ from .common import (
     bls_key_dir,
     sync_dir_map,
     harmony_dir,
-    validator_config,
-    saved_wallet_pass_path
 )
 from .blockchain import (
     get_latest_header,
-    get_latest_headers,
-    get_all_validator_addresses,
-    get_validator_information
 )
 from .util import (
     input_with_print,
@@ -168,68 +162,3 @@ def has_bad_block(log_file_path):
     except (UnicodeDecodeError, IOError):
         log(f"{Typgpy.WARNING}WARNING: failed to read `{log_file_path}` to check for bad block{Typgpy.ENDC}")
     return False
-
-
-def check_and_activate(epos_status_msg):
-    """
-    Return True when attempted to activate, otherwise return False.
-    """
-    if "not eligible" in epos_status_msg or "not signing" in epos_status_msg:
-        log(f"{Typgpy.FAIL}Node not active, reactivating...{Typgpy.ENDC}")
-        curr_headers = get_latest_headers("http://localhost:9500/")
-        curr_epoch_shard = curr_headers['shard-chain-header']['epoch']
-        curr_epoch_beacon = curr_headers['beacon-chain-header']['epoch']
-        wait_for_node_response(node_config['endpoint'], tries=900, sleep=1, verbose=False)  # Try for 15 min
-        ref_epoch = get_latest_header(node_config['endpoint'])['epoch']
-        if curr_epoch_shard == ref_epoch and curr_epoch_beacon == ref_epoch:
-            try:
-                activate_validator()
-                return True
-            except (TimeoutError, ConnectionError, RuntimeError, subprocess.CalledProcessError) as e:
-                log(f"{Typgpy.FAIL}Unable to activate validator {validator_config['validator-addr']}"
-                    f"error {e}. Continuing...{Typgpy.ENDC}")
-                return False
-        else:
-            log(f"{Typgpy.WARNING}Node not synced, did NOT activate node.{Typgpy.ENDC}")
-            return False
-    return False
-
-
-def deactivate_validator():
-    """
-    Assumption that endpoint is alive. Will throw error if not.
-    """
-    all_val = get_all_validator_addresses(node_config['endpoint'])
-    if validator_config["validator-addr"] in all_val:
-        val_chain_info = get_validator_information(validator_config["validator-addr"], node_config['endpoint'])
-        if "not eligible" not in val_chain_info['epos-status']:
-            log(f"{Typgpy.OKBLUE}Deactivating validator{Typgpy.ENDC}")
-            response = cli.single_call(
-                f"hmy staking edit-validator --validator-addr {validator_config['validator-addr']} "
-                f"--active false --node {node_config['endpoint']} "
-                f"--passphrase-file {saved_wallet_pass_path} --gas-price {validator_config['gas-price']} ")
-            log(f"{Typgpy.OKGREEN}Edit-validator response: {response}{Typgpy.ENDC}")
-        else:
-            log(f"{Typgpy.WARNING}Validator {validator_config['validator-addr']} is already deactivated!{Typgpy.ENDC}")
-    else:
-        log(f"{Typgpy.FAIL}Address {validator_config['validator-addr']} is not a validator!{Typgpy.ENDC}")
-
-
-def activate_validator():
-    """
-    Assumption that endpoint is alive. Will throw error if not.
-    """
-    all_val = get_all_validator_addresses(node_config['endpoint'])
-    if validator_config["validator-addr"] in all_val:
-        val_chain_info = get_validator_information(validator_config["validator-addr"], node_config['endpoint'])
-        if "not eligible" in val_chain_info['epos-status']:
-            log(f"{Typgpy.OKBLUE}Activating validator{Typgpy.ENDC}")
-            response = cli.single_call(
-                f"hmy staking edit-validator --validator-addr {validator_config['validator-addr']} "
-                f"--active true --node {node_config['endpoint']} "
-                f"--passphrase-file {saved_wallet_pass_path} --gas-price {validator_config['gas-price']} ")
-            log(f"{Typgpy.OKGREEN}Edit-validator response: {response}{Typgpy.ENDC}")
-        else:
-            log(f"{Typgpy.WARNING}Validator {validator_config['validator-addr']} is already active!{Typgpy.ENDC}")
-    else:
-        log(f"{Typgpy.FAIL}Address {validator_config['validator-addr']} is not a validator!{Typgpy.ENDC}")
